+++ CHANGELOG.md
@@ -296,3 +296,4 @@
 - 2024-02-17: Integ deal, mod, users, find email, +6
 - 2024-02-18: Wire README, frontier, discovery, mutate, +3
 - 2024-02-18: Pool frontier, users, bettercontact, follow up, +5
+- 2024-02-19: Pool discover, descend, task, mod, +4